code integer fixed

+++ database/migrations/2017_10_28_162759_create_transitions_table.php
@@ -15,7 +15,7 @@ class CreateTransitionsTable extends Migration
         Schema::create('transitions', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('fitness_id');
-            $table->float('value');
+            $table->bigInteger('value');
             $table->text('about');
             $table->integer('user_id');
             $table->timestamps();
